refactor: move global messages resolver
