+++ apps/web/src/schema/global-message/index.ts
@@ -1,0 +1,2 @@
+export * from './global-message.types'
+export * from './global-message.resolvers'
