+++ apps/web/src/schema/message/message.resolvers.ts
@@ -1,150 +1,16 @@
 /* eslint-disable no-console */
-import {
-  Resolver,
-  Query,
-  Mutation,
-  Ctx,
-  Arg,
-  ID,
-  Directive,
-} from 'type-graphql';
+import { Resolver, Query, Mutation, Ctx, Arg, ID } from 'type-graphql';
 
 import type { TContext } from '@/pages/api/graphql';
-import {
-  GlobalMessagesData,
-  SendGlobalMessage,
-  SendGlobalMessageInput,
-  SendMessageInput,
-  MessagesData,
-  GlobalMessage,
-} from './message.types';
+
 import { ErrorResponse } from '../types';
+import { SendMessageInput, MessagesData } from './message.types';
 
 @Resolver()
 export class MessageResolver {
   @Query(() => ErrorResponse)
   hello(): ErrorResponse {
     return { error: 'hi' };
-  }
-
-  @Directive('@cacheControl(maxAge: 240)')
-  @Query(() => GlobalMessagesData, { nullable: true })
-  async getGlobalMessages(
-    @Arg('cursorId', () => ID, { nullable: true }) cursorId: string,
-    @Ctx() { prisma }: TContext
-  ): Promise<GlobalMessagesData | null> {
-    try {
-      const messages = await prisma.globalMessage.findMany({
-        orderBy: { updatedAt: 'desc' },
-        take: 10,
-        include: {
-          user: {
-            select: {
-              id: true,
-              username: true,
-              image: true,
-            },
-          },
-        },
-        ...(cursorId && {
-          skip: 1,
-          cursor: {
-            id: cursorId,
-          },
-        }),
-      });
-
-      if (messages.length === 0) {
-        return {
-          data: [],
-          cursorId: null,
-        };
-      }
-
-      return {
-        data: messages,
-        cursorId: messages[messages.length - 1].id,
-      };
-    } catch (err) {
-      console.error(err);
-      throw err;
-    }
-  }
-
-  @Directive('@cacheControl(maxAge: 60)')
-  @Mutation(() => SendGlobalMessage)
-  async sendGlobalMessage(
-    @Arg('input', () => SendGlobalMessageInput)
-    { content, isAnonymous }: SendGlobalMessageInput,
-    @Ctx() { prisma, id }: TContext
-  ): Promise<SendGlobalMessage> {
-    let latestMessage: Omit<GlobalMessage, 'user'> | null;
-
-    try {
-      latestMessage = await prisma.globalMessage.findFirst({
-        where: { userId: id },
-        orderBy: { updatedAt: 'desc' },
-      });
-
-      if (latestMessage?.updatedAt) {
-        const diff = new Date().getTime() - latestMessage.updatedAt.getTime();
-
-        if (diff < 1000 * 60 && process.env.NODE_ENV !== 'development') {
-          return {
-            error: 'You can only send a message once every 1 minute.',
-          };
-        }
-      }
-    } catch (err) {
-      console.error(err);
-      throw err;
-    }
-
-    try {
-      let message: GlobalMessage;
-
-      if (!latestMessage) {
-        message = await prisma.globalMessage.create({
-          data: {
-            content,
-            isAnonymous,
-            user: id ? { connect: { id } } : undefined,
-          },
-          include: {
-            user: {
-              select: {
-                id: true,
-                username: true,
-                image: true,
-              },
-            },
-          },
-        });
-      } else {
-        message = await prisma.globalMessage.update({
-          where: { id: latestMessage?.id },
-          data: {
-            content,
-            isAnonymous,
-            user: id ? { connect: { id } } : undefined,
-          },
-          include: {
-            user: {
-              select: {
-                id: true,
-                username: true,
-                image: true,
-              },
-            },
-          },
-        });
-      }
-
-      return { data: message };
-    } catch (err) {
-      console.error(err);
-      throw err;
-    }
   }
 
   @Query(() => MessagesData, { nullable: true })
